# docs/swagger.yaml
basePath: /api/v1
definitions:
  httpapi.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  httpapi.IDResponse:
    properties:
      id:
        type: string
    type: object
  httpapi.StatusResponse:
    properties:
      status:
        type: string
    type: object
  httpapi.SubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  httpapi.SubscriptionResponse:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  httpapi.TotalResponse:
    properties:
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing user subscriptions and calculating totals.
  title: Subscriptions Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: List subscriptions with optional filters.
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription record.
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Sum of subscription costs for the specified period with optional
        filters.
      parameters:
      - description: Start period (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: End period (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.TotalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Calculate total subscriptions cost
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"